requirements.txt
python-multipart
fastapi~=0.111.0
lxml
beautifulsoup4~=4.12.3
pydantic==2.0.3
motor~=3.4.0
dnspython
python-decouple
pymongo~=4.7.3
openpyxl~=3.1.4
openpyxl-image-loader
geopandas~=0.14.4
shapely~=2.0.4

